[Decomp][Player] Cohesion split: extract enchantment-apply methods to PlayerEnchant.cpp
